Move media to src assets
