+++ src/assets/json/projects.json
@@ -5,11 +5,11 @@
         "name": "Sin designer",
         "toDisplayOnlyOne": {
           "img": "sin1",
-          "video": "video1"
+          "video": "sinVideo1"
         },
 
         "images": ["sin1", "sin2", "sin3", "sin4", "sin5"],
-        "video": ["video1", "video2", "video3"]
+        "video": ["sinVideo1", "sinVideo2", "sinVideo3"]
       }
     ]
   }
